Вынесена логика взятия курсов в отдельный файл
